refactor: organized src/lexer
