+++ src/lexer/lexer.rs
@@ -1,88 +1,4 @@
-use std::fmt;
-
-#[derive(Clone, PartialEq, Eq)]
-pub struct Position {
-    line: usize,
-    col: usize,
-    index: usize,
-}
-
-impl Position {
-    pub fn new() -> Self {
-        Self {
-            line: 0,
-            col: 0,
-            index: 0,
-        }
-    }
-}
-
-impl fmt::Debug for Position {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "(line: {}, col: {}, index: {})",
-            self.line, self.col, self.index
-        )
-    }
-}
-
-#[derive(Clone, PartialEq, Eq)]
-pub struct Token {
-    pub ttype: TokenType,
-    pub literal: String,
-    pub position: Position,
-}
-
-impl fmt::Debug for Token {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "[Type: {:?}, Literal: {}, Position: {:?}]",
-            self.ttype, self.literal, self.position
-        )
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum TokenType {
-    Keyword(KeywordType),
-    Ident,
-    Number,
-    Semicolon,
-    Mul,
-    Add,
-    Sub,
-    Div,
-    Assign,
-    NotEq,
-    Colon,
-    Lt,
-    Gt,
-    Eq,
-    Bang,
-    LParen,
-    RParen,
-    LBrace,
-    RBrace,
-    LBracket,
-    RBracket,
-    Comma,
-    Period,
-    String,
-    Eof,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum KeywordType {
-    Let,
-    Return,
-    True,
-    False,
-    If,
-    Else,
-    Fn,
-}
+use super::{KeywordType, Position, Token, TokenType};
 
 pub struct Lexer {
     src: String,
